legacy_query_result: move legacy helpers from session.rs
IntoTypedRows and TypedRowIter, being helper items for
LegacyQueryResult, are moved to legacy_query_result.rs from session.rs
not to clutter session.rs. They are still re-exported in session.rs not
to break compatibility.
